+++ v2/handlers.go
@@ -6,6 +6,8 @@ import (
 	"errors"
 	"fmt"
 
+	kadt "github.com/libp2p/go-libp2p-kad-dht/v2/kadt"
+
 	ds "github.com/ipfs/go-datastore"
 	record "github.com/libp2p/go-libp2p-record"
 	recpb "github.com/libp2p/go-libp2p-record/pb"
@@ -41,7 +43,7 @@ func (d *DHT) handleFindPeer(ctx context.Context, remote peer.ID, req *pb.Messag
 	}
 
 	// gather closer peers that we know
-	resp.CloserPeers = d.closerPeers(ctx, remote, nodeID(target).Key())
+	resp.CloserPeers = d.closerPeers(ctx, remote, kadt.PeerID(target).Key())
 
 	// if we happen to know the target peers addresses (e.g., although we are
 	// far away in the keyspace), we add the peer to the result set. This means
@@ -242,7 +244,7 @@ func (d *DHT) closerPeers(ctx context.Context, remote peer.ID, target key.Key256
 	// pre-allocated the result set slice.
 	filtered := make([]*pb.Message_Peer, 0, len(peers))
 	for _, p := range peers {
-		pid := peer.ID(p.(nodeID)) // TODO: type cast
+		pid := peer.ID(p.(kadt.PeerID)) // TODO: type cast
 
 		// check for own peer ID
 		if pid == d.host.ID() {
