+++ v2/handlers_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/ipfs/boxo/path"
 	ds "github.com/ipfs/go-datastore"
 	"github.com/libp2p/go-libp2p"
+	"github.com/libp2p/go-libp2p-kad-dht/v2/kadt"
 	pb "github.com/libp2p/go-libp2p-kad-dht/v2/pb"
 	record "github.com/libp2p/go-libp2p-record"
 	recpb "github.com/libp2p/go-libp2p-record/pb"
@@ -81,7 +82,7 @@ func fillRoutingTable(t testing.TB, d *DHT) {
 		pid := newPeerID(t)
 
 		// add peer to routing table
-		d.rt.AddNode(nodeID(pid))
+		d.rt.AddNode(kadt.PeerID(pid))
 
 		// craft network address for peer
 		a, err := ma.NewMultiaddr(fmt.Sprintf("/ip4/127.0.0.1/tcp/%d", 2000+i))
@@ -122,7 +123,7 @@ func BenchmarkDHT_handleFindPeer(b *testing.B) {
 		pid := newPeerID(b)
 
 		// add peer to routing table
-		d.rt.AddNode(nodeID(pid))
+		d.rt.AddNode(kadt.PeerID(pid))
 
 		// keep track of peer
 		peers = append(peers, pid)
@@ -174,7 +175,7 @@ func TestDHT_handleFindPeer_happy_path(t *testing.T) {
 		// closer peers. This means we can't assert for exactly 20 closer peers
 		// below.
 		if i > 0 {
-			d.rt.AddNode(nodeID(pid))
+			d.rt.AddNode(kadt.PeerID(pid))
 		}
 
 		// keep track of peer
@@ -208,7 +209,7 @@ func TestDHT_handleFindPeer_self_in_routing_table(t *testing.T) {
 	// a case that shouldn't happen
 	d := newTestDHT(t)
 
-	d.rt.AddNode(nodeID(d.host.ID()))
+	d.rt.AddNode(kadt.PeerID(d.host.ID()))
 
 	req := &pb.Message{
 		Type: pb.Message_FIND_NODE,
@@ -253,7 +254,7 @@ func TestDHT_handleFindPeer_unknown_addresses_but_in_routing_table(t *testing.T)
 		pid := newPeerID(t)
 
 		// add peer to routing table
-		d.rt.AddNode(nodeID(pid))
+		d.rt.AddNode(kadt.PeerID(pid))
 
 		// keep track of peer
 		peers[i] = pid
@@ -322,7 +323,7 @@ func TestDHT_handleFindPeer_request_for_self(t *testing.T) {
 		pid := newPeerID(t)
 
 		// add peer to routing table
-		d.rt.AddNode(nodeID(pid))
+		d.rt.AddNode(kadt.PeerID(pid))
 
 		// keep track of peer
 		peers[i] = pid
@@ -378,7 +379,7 @@ func TestDHT_handleFindPeer_request_for_known_but_far_peer(t *testing.T) {
 		// don't add first peer to routing table -> the one we're asking for
 		// don't add second peer -> the one that's requesting
 		if i > 1 {
-			d.rt.AddNode(nodeID(pid))
+			d.rt.AddNode(kadt.PeerID(pid))
 		}
 	}
