+++ v2/pb/Makefile
@@ -1,11 +1,15 @@
-all: clone reset
-	protoc --go_out=./ --go_opt=Mgithub.com/libp2p/go-libp2p-record/pb/record.proto=github.com/libp2p/go-libp2p-record/pb ./dht.proto
+all: clone reset build clean
+
+build:
+	protoc --go_out=./ --go_opt=Mgithub.com/libp2p/go-libp2p-record/pb/record.proto=github.com/libp2p/go-libp2p-record/pb ./msg.proto
 
 clone:
-	git clone --depth=1 git@github.com:libp2p/go-libp2p-record.git github.com/libp2p/go-libp2p-record || true
+	git clone --depth 1 --branch v0.2.0 git@github.com:libp2p/go-libp2p-record.git github.com/libp2p/go-libp2p-record || true
 
 reset:
-	rm dht.pb.go
+	rm msg.pb.go || true
 
 clean:
 	rm -rf github.com
+
+.PHONY: all build clone reset clean
